+++ samples/DemoApp/Properties/Resources.Designer.cs
@@ -19,7 +19,7 @@ namespace DemoApp.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "18.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Resources {
@@ -286,6 +286,33 @@ namespace DemoApp.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Map Required Properties Exception.
+        /// </summary>
+        internal static string Home_Button_MapPropertiesDemoWithoutProperties {
+            get {
+                return ResourceManager.GetString("Home_Button_MapPropertiesDemoWithoutProperties", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Map Properties demo.
+        /// </summary>
+        internal static string Home_Button_MapPropertiesDemoWithProperties {
+            get {
+                return ResourceManager.GetString("Home_Button_MapPropertiesDemoWithProperties", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Page Visibility Events demo.
+        /// </summary>
+        internal static string Home_Button_PageVisibilityEventsDemo {
+            get {
+                return ResourceManager.GetString("Home_Button_PageVisibilityEventsDemo", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Start Tabbed Page on Beta tab.
         /// </summary>
         internal static string Home_Button_StartTabbedPageOnBetaTab {
@@ -304,6 +331,15 @@ namespace DemoApp.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Events:.
+        /// </summary>
+        internal static string Home_SubHeading_Events {
+            get {
+                return ResourceManager.GetString("Home_SubHeading_Events", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Flyout navigation:.
         /// </summary>
         internal static string Home_SubHeading_FlyoutNavigation {
@@ -318,6 +354,15 @@ namespace DemoApp.Properties {
         internal static string Home_SubHeading_ModalNavigation {
             get {
                 return ResourceManager.GetString("Home_SubHeading_ModalNavigation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Parameter Passing:.
+        /// </summary>
+        internal static string Home_SubHeading_ParameterPassing {
+            get {
+                return ResourceManager.GetString("Home_SubHeading_ParameterPassing", resourceCulture);
             }
         }
         
@@ -426,6 +471,33 @@ namespace DemoApp.Properties {
         internal static string Login_Validation_RequiredUsername {
             get {
                 return ResourceManager.GetString("Login_Validation_RequiredUsername", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Map Properties.
+        /// </summary>
+        internal static string MapProperties_Title {
+            get {
+                return ResourceManager.GetString("MapProperties_Title", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to I show OnAppearing and hide OnDisappearing.
+        /// </summary>
+        internal static string PageVisibilityEvent_ShowOnAppearing {
+            get {
+                return ResourceManager.GetString("PageVisibilityEvent_ShowOnAppearing", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Page Visibility Events.
+        /// </summary>
+        internal static string PageVisibilityEvent_Title {
+            get {
+                return ResourceManager.GetString("PageVisibilityEvent_Title", resourceCulture);
             }
         }