+++ samples/DemoApp/ViewModels/Flyouts/ContactsViewModel.cs
@@ -9,7 +9,7 @@ public partial class ContactsViewModel : BaseViewModel
     #region Properties
 
     [ObservableProperty]
-    private List<string> contactNames = new List<string>()
+    public partial List<string> ContactNames { get; set; } = new List<string>()
         {
             "Ronan",
             "Claire",