+++ samples/DemoApp/ViewModels/LoginViewModel.cs
@@ -19,10 +19,10 @@ public partial class LoginViewModel : BaseViewModel
     #region Properties
 
     [ObservableProperty]
-    private string username;
+    public partial string Username { get; set; }
 
     [ObservableProperty]
-    private string password;
+    public partial string Password { get; set; }
 
     #endregion Properties