+++ samples/DemoApp/Views/MapPropertiesPage.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace DemoApp.Views;
+
+public partial class MapPropertiesPage : ContentPage
+{
+    public MapPropertiesPage()
+    {
+        InitializeComponent();
+    }
+}