+++ samples/DemoApp/Views/PageVisibilityEventPage.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace DemoApp.Views;
+
+public partial class PageVisibilityEventPage : ContentPage
+{
+    public PageVisibilityEventPage()
+    {
+        InitializeComponent();
+    }
+}