+++ src/Burkus.Mvvm.Maui/Services/NavigationService.cs
@@ -1,7 +1,12 @@
-﻿namespace Burkus.Mvvm.Maui;
+﻿using Microsoft.Extensions.Logging;
 
-internal class NavigationService : INavigationService
+namespace Burkus.Mvvm.Maui;
+
+internal class NavigationService(ILogger<NavigationService> logger)
+    : INavigationService
 {
+    internal ILogger<NavigationService>? Logger { get; } = logger;
+
     #region Core navigation methods
 
     public async Task Push<T>() where T : Page
@@ -13,15 +18,15 @@ internal class NavigationService : INavigationService
     {
         await HandleNavigation<T>(async () =>
             {
-                var pageToNavigateTo = ServiceResolver.Resolve<T>();
+                var pageToNavigateTo = GetPageToNavigateTo<T>();
 
                 if (navigationParameters.UseModalNavigation)
                 {
-                    await Application.Current.MainPage.Navigation.PushModalAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
+                    await MauiPageUtility.GetFirstWindow().Page.Navigation.PushModalAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
                 }
                 else
                 {
-                    await Application.Current.MainPage.Navigation.PushAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
+                    await MauiPageUtility.GetFirstWindow().Page.Navigation.PushAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
                 }
             },
             navigationParameters);
@@ -38,11 +43,11 @@ internal class NavigationService : INavigationService
             {
                 if (navigationParameters.UseModalNavigation)
                 {
-                    _ = await Application.Current.MainPage.Navigation.PopModalAsync(navigationParameters.UseAnimatedNavigation);
+                    _ = await MauiPageUtility.GetFirstWindow().Page.Navigation.PopModalAsync(navigationParameters.UseAnimatedNavigation);
                 }
                 else
                 {
-                    _ = await Application.Current.MainPage.Navigation.PopAsync(navigationParameters.UseAnimatedNavigation);
+                    _ = await MauiPageUtility.GetFirstWindow().Page.Navigation.PopAsync(navigationParameters.UseAnimatedNavigation);
                 }
             },
             navigationParameters);
@@ -57,7 +62,7 @@ internal class NavigationService : INavigationService
     {
         await HandleNavigation<Page>(async () =>
             {
-                await Application.Current.MainPage.Navigation.PopToRootAsync(navigationParameters.UseAnimatedNavigation);
+                await MauiPageUtility.GetFirstWindow().Page.Navigation.PopToRootAsync(navigationParameters.UseAnimatedNavigation);
             },
             navigationParameters);
     }
@@ -75,20 +80,22 @@ internal class NavigationService : INavigationService
     {
         await HandleNavigation<Page>(async () =>
             {
+                var window = MauiPageUtility.GetFirstWindow();
+
                 if (navigationParameters.UseModalNavigation)
                 {
-                    _ = await Application.Current.MainPage.Navigation.PopModalAsync(navigationParameters.UseAnimatedNavigation);
+                    _ = await window.Page.Navigation.PopModalAsync(navigationParameters.UseAnimatedNavigation);
                 }
                 else
                 {
-                    if (Application.Current.MainPage.Navigation.NavigationStack.Count <= 1)
+                    if (window.Page.Navigation.NavigationStack.Count <= 1)
                     {
                         // quit the app as this page is the last one
-                        Application.Current.Quit();
+                        Application.Current?.Quit();
                     }
                     else
                     {
-                        _ = await Application.Current.MainPage.Navigation.PopAsync(navigationParameters.UseAnimatedNavigation);
+                        _ = await window.Page.Navigation.PopAsync(navigationParameters.UseAnimatedNavigation);
                     }
                 }
             },
@@ -106,8 +113,8 @@ internal class NavigationService : INavigationService
     {
         await HandleNavigation<Page>(async () =>
             {
-                var navigation = Application.Current.MainPage.Navigation;
-                var pageToNavigateTo = ServiceResolver.Resolve<T>();
+                var navigation = MauiPageUtility.GetFirstWindow().Page.Navigation;
+                var pageToNavigateTo = GetPageToNavigateTo<T>();
 
                 navigation.InsertPageBefore(pageToNavigateTo, navigation.NavigationStack.Last());
                 await navigation.PopAsync(navigationParameters.UseAnimatedNavigation);
@@ -126,8 +133,8 @@ internal class NavigationService : INavigationService
     {
         await HandleNavigation<Page>(async () =>
             {
-                var navigation = Application.Current.MainPage.Navigation;
-                var pageToNavigateTo = ServiceResolver.Resolve<T>();
+                var navigation = MauiPageUtility.GetFirstWindow().Page.Navigation;
+                var pageToNavigateTo = GetPageToNavigateTo<T>();
 
                 if (navigation.NavigationStack.Count > 0)
                 {
@@ -156,10 +163,12 @@ internal class NavigationService : INavigationService
 
     public async Task Navigate(string uri, NavigationParameters navigationParameters)
     {
+        var window = MauiPageUtility.GetFirstWindow();
+
         // todo: how to consider modal navigation/animations/parameters etc. at each segment of the navigation
         // todo: need to consider what query parameter should and shouldn't be.
         // should they be only for passing simple variables (strings, bools etc.) rather than complex json objects
-        var navigation = Application.Current.MainPage.Navigation;
+        var navigation = window.Page.Navigation;
 
         var segments = UriUtility.GetUriSegments(uri);
         var instructions = segments.Select(UriUtility.ParseUriSegment)
@@ -201,16 +210,16 @@ internal class NavigationService : INavigationService
                 else
                 {
                     // push pages relatively onto the stack
-                    var pageToNavigateTo = ServiceResolver.Resolve(instructions[i].PageType) as Page;
+                    var pageToNavigateTo = GetPageToNavigateTo(instructions[i].PageType);
                     var pushNavigationParameters = instructions[i].QueryParameters.MergeNavigationParameters(navigationParameters);
 
                     if (pushNavigationParameters.UseModalNavigation)
                     {
-                        await Application.Current.MainPage.Navigation.PushModalAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
+                        await window.Page.Navigation.PushModalAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
                     }
                     else
                     {
-                        await Application.Current.MainPage.Navigation.PushAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
+                        await window.Page.Navigation.PushAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
                     }
 
                     await LifecycleEventUtility.TriggerOnNavigatedTo(
@@ -240,11 +249,11 @@ internal class NavigationService : INavigationService
             // pop final page
             if (popNavigationParameters.UseModalNavigation)
             {
-                _ = await Application.Current.MainPage.Navigation.PopModalAsync(navigationParameters.UseAnimatedNavigation);
+                _ = await window.Page.Navigation.PopModalAsync(navigationParameters.UseAnimatedNavigation);
             }
             else
             {
-                _ = await Application.Current.MainPage.Navigation.PopAsync(navigationParameters.UseAnimatedNavigation);
+                _ = await window.Page.Navigation.PopAsync(navigationParameters.UseAnimatedNavigation);
             }
 
             await LifecycleEventUtility.TriggerOnNavigatedFrom(
@@ -254,16 +263,16 @@ internal class NavigationService : INavigationService
         else
         {
             // push page relatively onto the stack
-            var pageToNavigateTo = ServiceResolver.Resolve(lastInstruction.PageType) as Page;
+            var pageToNavigateTo = GetPageToNavigateTo(lastInstruction.PageType);
             var pushNavigationParameters = lastInstruction.QueryParameters.MergeNavigationParameters(navigationParameters);
 
             if (pushNavigationParameters.UseModalNavigation)
             {
-                await Application.Current.MainPage.Navigation.PushModalAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
+                await window.Page.Navigation.PushModalAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
             }
             else
             {
-                await Application.Current.MainPage.Navigation.PushAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
+                await window.Page.Navigation.PushAsync(pageToNavigateTo, navigationParameters.UseAnimatedNavigation);
             }
 
             // remove all the pages that needed removed
@@ -287,88 +296,17 @@ internal class NavigationService : INavigationService
 
     #endregion URI navigation methods
 
-    #region Internal implementation
-
-    private async Task HandleNavigation<T>(Func<Task> navigationAction, NavigationParameters navigationParameters)
-        where T : Page
-    {
-        var fromBindingContext = MauiPageUtility.GetTopPageBindingContext();
-
-        await LifecycleEventUtility.TriggerOnNavigatingFrom(fromBindingContext, navigationParameters);
-
-        await navigationAction.Invoke();
-
-        await LifecycleEventUtility.TriggerOnNavigatedFrom(fromBindingContext, navigationParameters);
-
-        var toBindingContext = MauiPageUtility.GetTopPageBindingContext();
-        await LifecycleEventUtility.TriggerOnNavigatedTo(toBindingContext, navigationParameters);
-
-        // if the SelectTab parameter is used, we will switch tab
-        SelectTabFromParameters(navigationParameters);
-    }
-
-    private void SelectTabFromParameters(NavigationParameters navigationParameters)
-    {
-        if (string.IsNullOrWhiteSpace(navigationParameters.SelectTab))
-        {
-            // no tab to select
-            return;
-        }
-
-        var tabType = UriUtility.FindPageType(navigationParameters.SelectTab);
-        SelectTabWithType(tabType);
-    }
-
-    /// <summary>
-    /// This method allows the <see cref="SelectTab{T}()"/> to be called with reflection.
-    /// </summary>
-    private async void SelectTabWithType(Type tabType)
-    {
-        var selectTabMethod = GetType()
-            .GetMethod(nameof(SelectTab))
-            .MakeGenericMethod(tabType);
-        selectTabMethod.Invoke(this, null);
-    }
-
-    /// <summary>
-    /// This method searches for and tries to find a TabbedPage that is visible to the user.
-    /// </summary>
-    /// <param name="page">Page to search for a TabbedPage in</param>
-    /// <returns>A tabbeed page if found</returns>
-    private TabbedPage FindVisibleTabbedPage(Page page)
-    {
-        return page switch
-        {
-            TabbedPage tabbedPage => tabbedPage,
-            FlyoutPage { Detail: TabbedPage flyoutTabbedPage } => flyoutTabbedPage,
-            FlyoutPage { Detail: var detail } => GetTabbedPageFromNavigationPage(detail),
-            _ => GetTabbedPageFromNavigationPage(page)
-        };
-    }
-
-    private TabbedPage GetTabbedPageFromNavigationPage(Page page)
-    {
-        if (page is NavigationPage { CurrentPage: TabbedPage tabbedPage })
-        {
-            return tabbedPage;
-        }
-
-        return null;
-    }
-
-    #endregion Internal implementation
-
     #region Tab navigation methods
 
     public void SelectTab<T>()
         where T : Page
     {
         var topPage = MauiPageUtility.GetTopPage();
-        var tabbedPage = FindVisibleTabbedPage(topPage);
+        var tabbedPage = MauiPageUtility.FindVisibleTabbedPage(topPage);
 
         if (tabbedPage == null)
         {
-            // todo: warn about this in https://github.com/BurkusCat/Burkus.Mvvm.Maui/issues/17 ?
+            Logger?.LogWarning($"No visible {nameof(TabbedPage)} was found. {nameof(SelectTab)} must be called when a {nameof(TabbedPage)} is the top page.");
             return;
         }
 
@@ -411,11 +349,11 @@ internal class NavigationService : INavigationService
 
         if (flyoutPage == null)
         {
-            // todo: warn about this in https://github.com/BurkusCat/Burkus.Mvvm.Maui/issues/17 ?
+            Logger?.LogWarning($"No visible {nameof(FlyoutPage)} was found. {nameof(SwitchFlyoutDetail)} must be called when a {nameof(FlyoutPage)} is the top page.");
             return;
         }
 
-        var pageToNavigateTo = ServiceResolver.Resolve<T>();
+        var pageToNavigateTo = GetPageToNavigateTo<T>();
 
         // wrap the detail in a NavigationPage
         flyoutPage.Detail = new NavigationPage(pageToNavigateTo);
@@ -430,4 +368,75 @@ internal class NavigationService : INavigationService
     // a parameter? ToNavigationPage could be used so it could be used with URL navigation or the existing methods
 
     #endregion Flyout navigation methods
+
+    #region Internal implementation
+
+    private T GetPageToNavigateTo<T>()
+        where T: Page
+    {
+        var pageToNavigateTo = ServiceResolver.Resolve<T>();
+
+        LifecycleEventUtility.SubscribeToPageVisibilityEvents(pageToNavigateTo);
+
+        return pageToNavigateTo;
+    }
+
+    private Page? GetPageToNavigateTo(Type pageType)
+    {
+        var pageToNavigateTo = ServiceResolver.Resolve(pageType) as Page;
+
+        if (pageToNavigateTo != null)
+        {
+            LifecycleEventUtility.SubscribeToPageVisibilityEvents(pageToNavigateTo);
+        }
+        else
+        {
+            Logger?.LogWarning($"The type {pageType} could not be resolved to a {nameof(Page)}. It must be registered in the service container and inherit from {nameof(Page)}.");
+        }
+
+        return pageToNavigateTo;
+    }
+
+    private async Task HandleNavigation<T>(Func<Task> navigationAction, NavigationParameters navigationParameters)
+        where T : Page
+    {
+        var fromBindingContext = MauiPageUtility.GetTopPageBindingContext();
+
+        await LifecycleEventUtility.TriggerOnNavigatingFrom(fromBindingContext, navigationParameters);
+
+        await navigationAction.Invoke();
+
+        await LifecycleEventUtility.TriggerOnNavigatedFrom(fromBindingContext, navigationParameters);
+
+        var toBindingContext = MauiPageUtility.GetTopPageBindingContext();
+        await LifecycleEventUtility.TriggerOnNavigatedTo(toBindingContext, navigationParameters);
+
+        // if the SelectTab parameter is used, we will switch tab
+        SelectTabFromParameters(navigationParameters);
+    }
+
+    private void SelectTabFromParameters(NavigationParameters navigationParameters)
+    {
+        if (string.IsNullOrWhiteSpace(navigationParameters.SelectTab))
+        {
+            // no tab to select
+            return;
+        }
+
+        var tabType = UriUtility.FindPageType(navigationParameters.SelectTab);
+        SelectTabWithType(tabType);
+    }
+
+    /// <summary>
+    /// This method allows the <see cref="SelectTab{T}()"/> to be called with reflection.
+    /// </summary>
+    private async void SelectTabWithType(Type tabType)
+    {
+        var selectTabMethod = GetType()
+            .GetMethod(nameof(SelectTab))
+            .MakeGenericMethod(tabType);
+        selectTabMethod.Invoke(this, null);
+    }
+
+    #endregion Internal implementation
 }