+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2026-02-28
+
+### Added
+- Workaround to fix window title on WinUI
+- `ILogger` warnings when navigation methods are not used correctly (
+- `IPageVisibilityEvents` interface
+- `MapNavigationParameterAttribute` to map navigation parameters to properties
+  - Parameters can be marked as required
+
+### Changed
+- TargetFramework is now `net10.0` (instead of `net8.0`)
+- Updated Demo app fields to use partial properties (as recommended by CommunityToolkit.Mvvm)
+
+### Fixed
+- WinUI source generator error
 
 ## [0.2.0] - 2023-11-14